+++ controllers/contacts/index.js
@@ -1,0 +1,13 @@
+const listContacts = require("./listContacts")
+const getContactById = require("./getContactById")
+const addContact = require("./addContact")
+const updateContact = require("./updateContact")
+const removeContact = require("./removeContact")
+
+module.exports = {
+    listContacts,
+    getContactById,
+    addContact,
+    updateContact,
+    removeContact
+}
