+++ controllers/contacts/listContacts.js
@@ -1,0 +1,13 @@
+const contacts = require("../../models/contacts")
+
+const listContacts = async (req, res, next) => {
+  try {
+    const result = await contacts.listContacts()
+    res.json(result)
+  }
+  catch (error) {
+    next(error)
+  }
+}
+
+module.exports = listContacts;
